Translate ANSWERED to ACTIVE for ICS updates
Ensure that when we're processing a call state answered, that we treat
it as ACTIVE instead of RINGING. We're running into situations where
Telecom will attempt to connect to BT audio when the call state is still
answered but because we send this update to the ICS as ringing, BT ICS
does not think that we are in a call. As a result, they will reject the
requests to connect SCO. Sending the active call state instead will
ensure that BT ICS gets the update to be notified that there is a call.

Bug: 404473378
Flag: EXEMPT minor bugfix
Test: atest TelecomUnitTests
Change-Id: Ibec47f4845878249e1a6b505dbce70770d8926b9

CRs-Fixed: 4145776
Change-Id: Ia862f71e57ab9184fe9a9de1c1d545a97bf1ec3e